split search function to module.
